fix: transfer FAQ data to external files, simulating API request
